Fix [#140] 카테고리에 연결된 task가 없을 경우 핸들링하는 코드 추가

+++ jaksim/src/main/java/org/sopt/jaksim/category/facade/CategoryTaskFacade.java
@@ -75,7 +75,9 @@ public class CategoryTaskFacade {
     public void deleteCategoryTaskAndTasks(Long categoryId) {
         // CategoryTask -> Task 순으로 삭제
         List<Long> taskIdList = categoryTaskService.deleteAndGetTaskIdList(categoryId);
-        taskService.delete(taskIdList);
+        if (taskIdList != null) { // 카테고리에 등록된 태스크가 있는 경우에만 delete
+            taskService.delete(taskIdList);
+        }
     }
